Remove constants file and move url prop to CSWService
