--- service-csw/src/main/java/fi/nls/oskari/csw/Constants.java
@@ -1,6 +1,0 @@
-package fi.nls.oskari.csw;
-
-public class Constants {
-
-    public static final String PROP_SERVICE_URL = "service.metadata.url";
-}
